+++ crates/bitwarden-auth/src/registration/mod.rs
@@ -3,14 +3,14 @@
 //! mechanisms to establish their cryptographic state and register with
 //! the Bitwarden server
 
-mod open_org_invite;
+mod open_org_invite_crypto;
 mod post_keys_for_jit_password_registration;
 mod post_keys_for_key_connector_registration;
 mod post_keys_for_tde_registration;
 mod post_keys_for_user_password_registration;
 mod registration_client;
 
-pub use open_org_invite::{
+pub use open_org_invite_crypto::{
     OpenOrgInvite, SealedOpenOrgInvite, SealedOpenOrgInviteData, SealedOpenOrgInviteDataError,
 };
 pub use post_keys_for_jit_password_registration::{
